pull out dungeon scene
