+++ assets/levels/001.json
@@ -61,7 +61,7 @@
                  "visible":true,
                  "width":0,
                  "x":43.4634974533107,
-                 "y":87.8324844368987
+                 "y":86.0215053763441
                 }],
          "opacity":1,
          "type":"objectgroup",
