+++ assets/levels/001.tmx
@@ -32,7 +32,7 @@
 </data>
  </layer>
  <objectgroup id="3" name="markers">
-  <object id="1" name="hero-start" x="43.4635" y="87.8325">
+  <object id="1" name="hero-start" x="43.4635" y="86.0215">
    <point/>
   </object>
  </objectgroup>
