+++ src/index.ts
@@ -1,7 +1,6 @@
 import Phaser from 'phaser';
 import { PhecsPlugin } from 'phecs';
-import { LevelCreator } from './level-creator';
-import { Depths } from './constants/depths';
+import { DungeonScene } from './scenes/dungeon-scene';
 
 class GameScene extends Phaser.Scene {
   phecs!: PhecsPlugin;
@@ -20,25 +19,7 @@ class GameScene extends Phaser.Scene {
     */
   }
 
-  preload() {
-    this.load.image('dungeon-spritesheet', 'assets/maps/dungeon-spritesheet.png');
-    this.load.tilemapTiledJSON('level-001', 'assets/levels/001.json');
-
-    this.load.spritesheet('hero', 'assets/characters/hero/spritesheet.png', { frameWidth: 32, frameHeight: 56 });
-    this.load.animation('hero-animations', 'assets/characters/hero/animations.json');
-  }
-
   create() {
-    this.cameras.main.setBackgroundColor(0xCCCCCC);
-
-    const levelCreator = new LevelCreator(this, 'level-001');
-    levelCreator.load();
-    levelCreator.createMap(100, 100);
-
-    const heroStartWorldCoordinates = levelCreator.getHeroStartWorldPosition();
-    const hero = this.add.sprite(heroStartWorldCoordinates.x, heroStartWorldCoordinates.y, 'hero');
-    hero.setDepth(Depths.hero);
-
     /*
     this.phecs.add.prefab('point', {}, 10, 20);
     this.phecs.add.prefab('point', {}, 100, 20);
@@ -52,7 +33,7 @@ class GameScene extends Phaser.Scene {
 const game = new Phaser.Game({
   width: 352,
   height: 600,
-  scene: [GameScene],
+  scene: [DungeonScene],
   plugins: {
     scene: [
       {
