+++ src/scenes/dungeon-scene.ts
@@ -1,0 +1,24 @@
+import { Depths } from "../constants/depths";
+import { LevelCreator } from "../level-creator";
+
+export class DungeonScene extends Phaser.Scene {
+  preload() {
+    this.load.image('dungeon-spritesheet', 'assets/maps/dungeon-spritesheet.png');
+    this.load.tilemapTiledJSON('level-001', 'assets/levels/001.json');
+
+    this.load.spritesheet('hero', 'assets/characters/hero/spritesheet.png', { frameWidth: 32, frameHeight: 56 });
+    this.load.animation('hero-animations', 'assets/characters/hero/animations.json');
+  }
+
+  create() {
+    this.cameras.main.setBackgroundColor(0xCCCCCC);
+
+    const levelCreator = new LevelCreator(this, 'level-001');
+    levelCreator.load();
+    levelCreator.createMap(100, 100);
+
+    const heroStartWorldCoordinates = levelCreator.getHeroStartWorldPosition();
+    const hero = this.add.sprite(heroStartWorldCoordinates.x, heroStartWorldCoordinates.y, 'hero');
+    hero.setDepth(Depths.hero);
+  }
+}
